Module context::context

Provides the Context and Transfer types for saving and restoring the current state of execution.

See the Context struct for more information.

Structs

Context
A Context stores a ContextFn's state of execution, for it to be resumed later.

Transfer
Contains the previously active Context and the data passed to resume the current one and is used as the return value by Context::resume() and Context::resume_ontop()

Type Definitions

ContextFn
Functions of this signature are used as the entry point for a new Context.

ResumeOntopFn
Functions of this signature are used as the callback while resuming ontop of a Context.