Struct context::context::Transfer

pub struct Transfer {
    pub context: Context,
    pub data: usize,
}

Contains the previously active Context and the data passed to resume the current one and is used as the return value by Context::resume() and Context::resume_ontop()

Fields

context

The previously executed Context which yielded to resume the current one.

data

The data which was passed to Context::resume() or Context::resume_ontop() to resume the current Context.

Methods

impl Transfer

fn new(context: Context, data: usize) -> Transfer

Returns a new Transfer struct with the members set to their respective arguments.