Struct context::stack::ProtectedFixedSizeStack

pub struct ProtectedFixedSizeStack(_);

A more secure, but slightly slower version of FixedSizeStack.

Allocates stack space using virtual memory, whose pages will only be mapped to physical memory if they are used.

The additional guard page is made protected and inaccessible. Now if a stack overflow occurs it should (hopefully) hit this guard page and cause a segmentation fault instead letting the memory being overwritten silently.

As a general rule it is recommended to use this struct to create stack memory.

Methods

impl ProtectedFixedSizeStack

fn new(size: usize) -> Result<ProtectedFixedSizeStack, StackError>

Allocates a new stack of at least size bytes + one additional guard page.

size is rounded up to a multiple of the size of a memory page and does not include the size of the guard page itself.