Struct context::stack::Stack

pub struct Stack {
    // some fields omitted
}

Represents any kind of stack memory.

FixedSizeStack as well as ProtectedFixedSizeStack can be used to allocate actual stack space.

Methods

impl Stack

fn new(top: *mut c_void, bottom: *mut c_void) -> Stack

Creates a (non-owning) representation of some stack memory.

fn top(&self) -> *mut c_void

Returns the top of the stack from which on it grows downwards towards bottom().

fn bottom(&self) -> *mut c_void

Returns the bottom of the stack and thus it's end.

fn len(&self) -> usize

Returns the size of the stack between top() and bottom().

fn min_size() -> usize

Returns the minimal stack size allowed by the current platform.

fn max_size() -> usize

Returns the maximum stack size allowed by the current platform.

fn default_size() -> usize

Returns a implementation defined default stack size.

This value can vary greatly between platforms, but is usually only a couple memory pages in size and enough for most use-cases with little recursion. It's usually a better idea to specifiy an explicit stack size instead.