Struct context::context::Context

pub struct Context {
    // some fields omitted
}

Methods

impl Context

fn empty() -> Context

fn new<'a, F>(init: InitFn, arg: usize, start: Box<F>, stack: &mut Stack) -> Context where F: FnBox() + 'a

Create a new context that will resume execution by running start

The init function will be run with arg and the start procedure split up into code and env pointers. It is required that the init function never return.

FIXME: this is basically an awful the interface. The main reason for this is to reduce the number of allocations made when a green task is spawned as much as possible

fn init_with<'a, F>(&mut self, init: InitFn, arg: usize, start: Box<F>, stack: &mut Stack) where F: FnBox() + 'a

fn init_with_unboxed(&mut self, init: InitFn, arg: usize, stack: &mut Stack)

fn swap(out_context: &mut Context, in_context: &Context)

Switch contexts Suspend the current execution context and resume another by saving the registers values of the executing thread to a Context then loading the registers from a previously saved Context.

fn save(context: &mut Context)

Save the current context.

fn load(to_context: &Context)

Load the context and switch. This function will never return.

It is equivalent to Context::swap(&mut dummy_context, &to_context).