Struct context::stack::Stack

pub struct Stack {
    // some fields omitted
}

A task's stack. The name "Stack" is a vestige of segmented stacks.

Methods

impl Stack

fn new(size: usize) -> Result<Stack, StackError>

Allocate a new stack of size. If size = 0, this will fail. Use dummy_stack if you want a zero-sized stack.

unsafe fn dummy_stack() -> Stack

Create a 0-length stack which starts (and ends) at 0.

fn guard(&self) -> *const usize

Return the address of the guard page

fn start(&self) -> *const usize

Point to the low end of the allocated stack

fn end(&self) -> *const usize

Point one usize beyond the high end of the allocated stack